Adding module to setup tailscale
Not sure if setting up the authkey in age makes any sense. Authkeys are
shortlived, even reusable keys last at most three months. Additionaly it
is probably best to update the authkey on system rebuilds as well. If
the authkey is wrong the system hangs on boot as well, so additionally a
timeout should probably be added to the systemd service.

see:https://tailscale.com/blog/nixos-minecraft/